Major refactor to better adhere to SOLID principles
