+++ src/index.js
@@ -2,7 +2,6 @@ import "@picocss/pico";
 
 import getTaskForm from "./components/taskForm";
 import Task from "./logic/task";
-import tasksDisplay from "./components/tasks";
 
 import Header from "./components/Header";
 
@@ -10,29 +9,50 @@ const taskList = [];
 
 function removeTask(id) {
 	taskList.filter((currTask) => currTask.id !== id);
+	document.getElementById(id).remove();
 }
 
 function addTask(config) {
 	taskList.push(new Task(config));
-}
-
-function showTasks(parent, taskList) {
 	const taskListElem = document.getElementById("taskList");
-	if (taskListElem) parent.removeChild(taskListElem);
-	parent.appendChild(tasksDisplay(taskList, removeTask));
+
+	taskListElem.appendChild(getTaskElem(config));
 }
 
+function tasksDisplay() {
+	const taskListElem = document.createElement("section");
+	taskListElem.id = "taskList";
+	taskListElem.classList.add("container");
 
-//FOR TESTING vvvv
+	return taskListElem;
+}
 
-addTask({ title: "a", description: "b" });
-addTask({ title: "c", description: "d" });
+function getTaskElem({ title, description, id }) {
+	const task = document.createElement("details");
+	task.id = id;
+	const titleElem = document.createElement("summary");
+	titleElem.innerText = title;
 
-//FOR TESTING ^^^^
+	const descriptionElem = document.createElement("p");
+
+	descriptionElem.innerText = description;
+
+	const removeBtn = document.createElement("button");
+	removeBtn.innerText = "X";
+	removeBtn.onclick = (e) => {
+		removeTask(id);
+	};
+
+	titleElem.appendChild(removeBtn);
+
+	task.appendChild(titleElem);
+	task.appendChild(descriptionElem);
+	return task;
+}
 
 const component = () => {
 	const component = document.createElement("main");
-    component.classList.add("container")
+	component.classList.add("container");
 
 	const taskOnSubmit = (e) => {
 		e.preventDefault();
@@ -52,7 +72,6 @@ const component = () => {
 			dueDate: [dueDate, dueTime],
 		});
 
-		showTasks(component, taskList);
 		console.log(taskList);
 	};
 
@@ -61,15 +80,32 @@ const component = () => {
 		main.appendChild(getTaskForm(taskOnSubmit));
 	};
 
-    const onAddProject = () => {
-        return "cum"
-    }
+	const onAddProject = () => {
+		return "cum";
+	};
 
 	component.appendChild(Header({ onAddTask, onAddProject }));
 
-	showTasks(component, taskList);
+	const taskListElem = tasksDisplay();
+
+	taskList
+		.filter((task) => task.type === "task")
+		.map((task) => getTaskElem(task))
+		.reduce((acc, task) => {
+			acc.appendChild(task);
+			return acc;
+		}, taskListElem);
+
+	component.appendChild(taskListElem);
 
 	return component;
 };
 
 document.body.appendChild(component());
+
+//FOR TESTING vvvv
+
+addTask({ title: "a", description: "b" });
+addTask({ title: "c", description: "d" });
+
+//FOR TESTING ^^^^
